refactor: extract bb::ipc transport runtime to /ipc-runtime/cpp/
Lift the UDS + MPSC-SHM server/client primitives out of barretenberg into
a new top-level package /ipc-runtime/, parallel to /ipc-codegen/. The
runtime was already self-contained (one stale throw_or_abort.hpp include
in shm_server.hpp, never called) so this is a pure relocation.

Layout:

  /ipc-runtime/
    bootstrap.sh           # delegates to barretenberg/cpp/'s build for now
    .rebuild_patterns
    cpp/
      CMakeLists.txt       # ipc_runtime static lib + ipc_runtime_tests
      ipc_runtime/         # public headers + impl, matches include prefix
        ipc_{client,server}.{hpp,cpp}
        socket_{client,server}.{hpp,cpp}
        mpsc_shm_{client,server}.hpp
        shm_{client,server,common}.hpp
        shm.test.cpp
        shm/{futex,mpsc_shm,spsc_shm,utilities}.{hpp,cpp}

barretenberg/cpp/src/CMakeLists.txt picks the lib up via
add_subdirectory(${CMAKE_SOURCE_DIR}/../../ipc-runtime/cpp ...) so
in-tree consumers (bb, bb-avm, aztec-wsdb, nodejs_module, api,
ipc_bench) just retarget `#include "barretenberg/ipc/..."` to
`#include "ipc_runtime/..."` and `target_link_libraries(... ipc)` to
`... ipc_runtime`.

The shm.test.cpp gtest binary is now ipc_runtime_tests, runs as part of
barretenberg's CTest discovery, all 2 tests pass.

The next step (truly standalone build via /ipc-runtime/bootstrap.sh,
without barretenberg's CMake tree) is a follow-up: the bootstrap.sh
currently delegates to barretenberg's build, and ipc-runtime/cpp/
CMakeLists.txt still uses barretenberg_module_with_sources for gtest +
msgpack convenience. Files + namespace boundary are correct; the build
plumbing tidies up later.
