+++ barretenberg/cpp/src/barretenberg/bb/CMakeLists.txt
@@ -23,7 +23,7 @@ if (NOT(FUZZING))
         target_link_libraries(bb PRIVATE avm_transpiler)
     endif()
     if(NOT WASM AND NOT BB_LITE)
-        target_link_libraries(bb PRIVATE ipc)
+        target_link_libraries(bb PRIVATE ipc_runtime)
     endif()
     if(ENABLE_STACKTRACES)
         target_link_libraries(
@@ -63,7 +63,7 @@ if (NOT(FUZZING))
             target_link_libraries(bb-avm PRIVATE avm_transpiler)
         endif()
         if(NOT WASM AND NOT BB_LITE)
-            target_link_libraries(bb-avm PRIVATE ipc)
+            target_link_libraries(bb-avm PRIVATE ipc_runtime)
         endif()
         if(ENABLE_STACKTRACES)
             target_link_libraries(
