+++ barretenberg/cpp/src/barretenberg/nodejs_module/CMakeLists.txt
@@ -27,7 +27,7 @@ string(REGEX REPLACE "[\r\n\"]" "" NODE_API_HEADERS_DIR ${NODE_API_HEADERS_DIR})
 add_library(nodejs_module SHARED ${SOURCE_FILES})
 set_target_properties(nodejs_module PROPERTIES PREFIX "" SUFFIX ".node")
 target_include_directories(nodejs_module PRIVATE ${NODE_API_HEADERS_DIR} ${NODE_ADDON_API_DIR})
-target_link_libraries(nodejs_module PRIVATE world_state ipc vm2_sim)
+target_link_libraries(nodejs_module PRIVATE world_state ipc_runtime vm2_sim)
 
 # On macOS, Node.js N-API symbols are provided by the runtime, not at link time
 if(CMAKE_SYSTEM_NAME STREQUAL "Darwin")
