+++ barretenberg/cpp/src/barretenberg/nodejs_module/msgpack_client/msgpack_client_async.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "barretenberg/ipc/ipc_client.hpp"
+#include "ipc_runtime/ipc_client.hpp"
 #include "napi.h"
 #include <atomic>
 #include <mutex>
