+++ ipc-runtime/.rebuild_patterns
@@ -1,0 +1,3 @@
+^ipc-runtime/cpp/.*\.(cpp|hpp)$
+^ipc-runtime/cpp/CMakeLists\.txt$
+^ipc-runtime/bootstrap\.sh$
