+++ ipc-runtime/cpp/ipc_runtime/grind_ipc.sh
