+++ ipc-runtime/cpp/ipc_runtime/shm/utilities.hpp
